requirements.txt
python-dotenv == 0.19.0
requests ~= 2.26.0
nextcord~=2.0.0a3
asyncpraw ~= 7.4.0